requirements.txt
# Core dependencies for the InsightForge app
streamlit==1.50.0
pandas==2.3.3
matplotlib==3.10.7
seaborn==0.13.3
scikit-learn==1.5.0  # or match your tested version
faiss-cpu==1.12.0
openai==1.109.1
langchain==0.3.27
langchain-openai==0.3.35
langchain-community==0.3.31
ragas==0.3.6
datasets==4.2.0
evaluate==0.4.6
python-dotenv==1.1.1


# pip install -r requirements.txt